OMAP: HDMI: Correction in HDMI PLL program sequence
Current code sets PLL_GO to 1 and the waits for the PLL_GO to
become 1. This is not required and instead we should wait for
PLL_GO to become zero. As per TRM once PLL operation is over
PLL_GO bit will be reset to zero. After waiting for PLL_GO to
become zero PLL_STATUS register is read to check the PLL LOCK
status.

Change-Id: I6ce0d0839d761ed1bc46cc1ee5d66ace76ea5023
Author: Muralidhar Dixit <murali.dixit@ti.com>
Signed-off-by: Vadym Shovkoplias <shovkoplias.vadym@ti.com>