OMAP4: HDMI: Fix HDMI_TXPHY_TX_CONTROL.FREQOUT programming
According to TRM, the HDMI_TXPHY_TX_CONTROL.FREQOUT bit has to be programmed
based on the clock frequency. But this bit is always programmed as 1 in the HDMI
driver irrespective of the clock frequency values. Fix this and program
HDMI_TXPHY_TX_CONTROL.FREQOUT based on the clock frequency.

Change-Id: I38c8753408a4e74f95a880b3c9e65fb68aec8813
Author: TAKAHASHI, Masato <m_takahashi@ti.com>
Author: Varadarajan, Charulatha <charu@ti.com>
Signed-off-by: Vadym Shovkoplias <shovkoplias.vadym@ti.com>